w3c-reports/12092014/gtk/runned-tests/eventsource/interfaces-expected.txt

CONSOLE MESSAGE: line 188: callback not yet supported
CONSOLE MESSAGE: Deprecated attempt to access property 'url' on a non-EventSource object.
CONSOLE MESSAGE: Deprecated attempt to access property 'withCredentials' on a non-EventSource object.
CONSOLE MESSAGE: Deprecated attempt to access property 'readyState' on a non-EventSource object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onopen' on a non-EventSource object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onmessage' on a non-EventSource object.
CONSOLE MESSAGE: Deprecated attempt to access property 'onerror' on a non-EventSource object.
EventSource IDL tests


FAIL EventSource interface: existence and properties of interface object assert_equals: prototype of window's property "EventSource" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
PASS EventSource interface object length 
FAIL EventSource interface: existence and properties of interface prototype object assert_false: EventSource.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: existence and properties of interface prototype object's "constructor" property assert_true: EventSource.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: attribute url assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: attribute withCredentials assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
PASS EventSource interface: constant CONNECTING on interface object 
PASS EventSource interface: constant CONNECTING on interface prototype object 
PASS EventSource interface: constant OPEN on interface object 
PASS EventSource interface: constant OPEN on interface prototype object 
PASS EventSource interface: constant CLOSED on interface object 
PASS EventSource interface: constant CLOSED on interface prototype object 
FAIL EventSource interface: attribute readyState assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: attribute onopen assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: attribute onmessage assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
FAIL EventSource interface: attribute onerror assert_throws: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1270:26
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
PASS EventSource interface: operation close() 
PASS EventSource must be primary interface of new EventSource("http://foo") 
FAIL Stringification of new EventSource("http://foo") undefined is not an object (evaluating 'this.array.members[this.base].has_stringifier')(stack: has_stringifier@http://localhost:8800/resources/idlharness.js:1662:38
http://localhost:8800/resources/idlharness.js:1550:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_primary_interface_of@http://localhost:8800/resources/idlharness.js:1545:9
test_object@http://localhost:8800/resources/idlharness.js:1495:35
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
onload@http://localhost:8800/eventsource/interfaces.html:55:15)
PASS EventSource interface: new EventSource("http://foo") must inherit property "url" with the proper type (0) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "withCredentials" with the proper type (1) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "CONNECTING" with the proper type (2) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "OPEN" with the proper type (3) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "CLOSED" with the proper type (4) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "readyState" with the proper type (5) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "onopen" with the proper type (6) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "onmessage" with the proper type (7) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "onerror" with the proper type (8) 
PASS EventSource interface: new EventSource("http://foo") must inherit property "close" with the proper type (9) 
